+++ Communication/Client/Logic/ResponseActions/IResponseActions.cs
@@ -1,6 +1,10 @@
-﻿namespace Communication.Client.Logic.ResponseActions
+﻿using System.Text.RegularExpressions;
+
+namespace Communication.Client.Logic.ResponseActions
 {
     public interface IResponseActions
     {
+        void OnSuccess(GroupCollection groups);
+        void OnFailed(GroupCollection groups);
     }
 }
